# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sellify API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: Search and browse active listings
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: city
          in: query
          schema: { type: string }
        - name: country
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: condition
          in: query
          schema: { type: string }
        - name: sellerId
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, price_asc, price_desc] }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Listing search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingSearchResult"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a listing draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /listings/slug/{slug}:
    get:
      operationId: getListingBySlug
      tags: [listings]
      summary: Get a public listing by slug (no auth, increments view count)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /listings/{id}/publish:
    post:
      operationId: publishListing
      tags: [listings]
      summary: Publish a draft listing (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}/sold:
    post:
      operationId: markListingSold
      tags: [listings]
      summary: Mark a listing as sold (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Sold listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}/favorite:
    post:
      operationId: toggleFavorite
      tags: [listings]
      summary: Toggle favorite on a listing (auth required)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Favorite state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteState"
  /listings/{id}/similar:
    get:
      operationId: getSimilarListings
      tags: [listings]
      summary: Similar active listings (same category, nearby)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Similar listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/{id}/stats:
    get:
      operationId: getListingStats
      tags: [listings]
      summary: Seller stats for a listing (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingStats"
...
  /home:
    get:
      operationId: getHomeFeed
      tags: [listings]
      summary: Home feed with recommended, newest and nearby listings
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: country
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Home feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeFeed"
...
  /sellers/{id}:
    get:
      operationId: getSeller
      tags: [sellers]
      summary: Public seller profile with active listings
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerPublic"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
    Category:
      type: object
      required: [id, slug, nameSv, nameEn, icon, listingCount]
      properties:
        id: { type: integer }
        slug: { type: string }
        nameSv: { type: string }
        nameEn: { type: string }
        icon: { type: string, description: "Lucide icon name" }
        listingCount: { type: integer }
    Listing:
      type: object
      required:
        [
          id,
          sellerId,
          title,
          description,
          status,
          price,
          currency,
          condition,
          images,
          city,
          country,
          slug,
          shipping,
          viewCount,
          favoriteCount,
          isFavorited,
          createdAt,
        ]
      properties:
        id: { type: integer }
        sellerId: { type: string }
        sellerName: { type: ["string", "null"] }
        sellerAvatarUrl: { type: ["string", "null"] }
        title: { type: string }
        description: { type: string }
        shortDescription: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryNameSv: { type: ["string", "null"] }
        categoryNameEn: { type: ["string", "null"] }
        brand: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        material: { type: ["string", "null"] }
        condition:
          type: string
          enum: [new, like_new, good, fair, worn]
        status:
          type: string
          enum: [draft, active, sold]
        price: { type: number }
        currency: { type: string }
        priceType:
          type: ["string", "null"]
          enum: [fixed, negotiable, null]
        city: { type: string }
        region: { type: ["string", "null"] }
        country: { type: string }
        postalCode: { type: ["string", "null"] }
        shipping:
          type: string
          enum: [pickup, ship, both]
        images:
          type: array
          items: { type: string }
          description: Ordered image URLs; first is the main image.
        slug: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        keywords:
          type: ["array", "null"]
          items: { type: string }
        specifications:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/Specification"
        viewCount: { type: integer }
        favoriteCount: { type: integer }
        isFavorited: { type: boolean }
        createdAt: { type: string }
        publishedAt: { type: ["string", "null"] }
        soldAt: { type: ["string", "null"] }
    Specification:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
    ListingInput:
      type: object
      required: [title, description, price, currency, condition, images, city, country, shipping]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        shortDescription: { type: string }
        categoryId: { type: ["integer", "null"] }
        brand: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        material: { type: ["string", "null"] }
        condition: { type: string, enum: [new, like_new, good, fair, worn] }
        price: { type: number }
        currency: { type: string }
        priceType: { type: string, enum: [fixed, negotiable] }
        city: { type: string }
        region: { type: ["string", "null"] }
        country: { type: string }
        postalCode: { type: ["string", "null"] }
        shipping: { type: string, enum: [pickup, ship, both] }
        images:
          type: array
          items: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        keywords:
          type: array
          items: { type: string }
        specifications:
          type: array
          items:
            $ref: "#/components/schemas/Specification"
        status: { type: string, enum: [draft, active] }
    ListingUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        shortDescription: { type: string }
        categoryId: { type: ["integer", "null"] }
        brand: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        material: { type: ["string", "null"] }
        condition: { type: string, enum: [new, like_new, good, fair, worn] }
        price: { type: number }
        currency: { type: string }
        priceType: { type: string, enum: [fixed, negotiable] }
        city: { type: string }
        region: { type: ["string", "null"] }
        country: { type: string }
        postalCode: { type: ["string", "null"] }
        shipping: { type: string, enum: [pickup, ship, both] }
        images:
          type: array
          items: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        keywords:
          type: array
          items: { type: string }
        specifications:
          type: array
          items:
            $ref: "#/components/schemas/Specification"
    ListingSearchResult:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total: { type: integer }
    FavoriteState:
      type: object
      required: [favorited, favoriteCount]
      properties:
        favorited: { type: boolean }
        favoriteCount: { type: integer }
    ListingStats:
      type: object
      required: [viewCount, favoriteCount, messageCount, daysSincePublished, suggestions]
      properties:
        viewCount: { type: integer }
        favoriteCount: { type: integer }
        messageCount: { type: integer }
        daysSincePublished: { type: integer }
        suggestions:
          type: array
          items: { type: string }
          description: Simple AI-style suggestions based on the stats.
...
    SellerPublic:
      type: object
      required: [id, displayName, memberSince, activeListingCount, soldListingCount, listings]
      properties:
        id: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        memberSince: { type: string }
        activeListingCount: { type: integer }
        soldListingCount: { type: integer }
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
    HomeFeed:
      type: object
      required: [newest, nearby, categories, totalActive]
      properties:
        newest:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        nearby:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        totalActive: { type: integer }
...